+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-inline-content/Interlinking/index.ts
@@ -1,2 +1,1 @@
 export * from './InterlinkingLinkInlineContent';
-export * from './InterlinkingSearchInlineContent';
