+++ src/frontend/apps/impress/src/features/docs/doc-export/mappingPDF.tsx
@@ -30,7 +30,6 @@ export const pdfDocsSchemaMappings: DocsExporterPDF['mappings'] = {
   },
   inlineContentMapping: {
     ...pdfDefaultSchemaMappings.inlineContentMapping,
-    interlinkingSearchInline: () => <></>,
     interlinkingLinkInline: inlineContentMappingInterlinkingLinkPDF,
   },
   styleMapping: {
